minor refactor and additional PR changes
