--- newfunc/.funcignore
@@ -1,5 +1,0 @@
-
-# Use the .funcignore file to exclude files which should not be
-# tracked in the image build. To instruct the system not to track
-# files in the image build, add the regex pattern or file information
-# to this file.
