--- newfunc/go.mod
@@ -1,3 +1,0 @@
-module function
-
-go 1.21
