+++ pkg/functions/client_int_test.go
@@ -454,7 +454,7 @@ func Handle(res http.ResponseWriter, req *http.Request) {
 // TestInt_Invoke_ServiceToService ensures that a Function can invoke another
 // service via localhost service discovery api provided by the Dapr sidecar.
 func TestInt_Invoke_ServiceToService(t *testing.T) {
-	t.Skip("TODO: dapr appears to be borked")
+	t.Skip("TODO: dapr appears to be borked") // https://github.com/knative/func/issues/3210
 	resetEnv()
 	var (
 		verbose = true
