--- testpy/.func/local.yaml
@@ -1,1 +1,0 @@
-{}
